requirements.txt
fastmcp>=0.4.1
google-generativeai==0.3.2
python-dotenv>=1.0.0
loguru>=0.7.0
openai>=1.0.0 